feat: custom web UI for login/logout, replacing browser's basic auth (#716)
The basic auth popup is not compatible with password managers. It also doesn't
allow us to add a logout button - the only way to end a session is to close the
browser.

This commit implements a normal HTML login form, which password managers can
fill. The server will not offer the basic auth popup for requests coming from
the web UI.

We also include a logout button, now that we have the ability to end a session.

The RPC endpoint will now return 204 No Content on empty requests. This allows
the web UI to test whether it's authenticated by making a no-op network call.

The login form must be exposed to unauthenticated users, and web/ must also be
in order for the client to detect the absence of auth and navigate to the login
page.

Javascript uses localStorage to store auth credentials, then attaches a basic
auth header to fetch requests. If the server responds with HTTP 401, the client
navigates to the login page.

We now redirect / and /transmission/ to /transmission/web/ prior to checking
auth, so the auth flow is always the same.

The esbuild config was adjusted to output new JS + CSS files for the login
page. The login page is styled with oklch colors, which are easy to tweak by
hand for brightness. They are broadly supported by browsers.o

The app can now detect the URL path of the app root, because dealing with
relative URLs in JS called from multiple pages was getting messy.